// types/strapi/product.ts
// Interface automatically generated by schemas-to-ts

import { Media } from "../common/schemas-to-ts/Media";
import { AdminPanelRelationPropertyModification } from "../common/schemas-to-ts/AdminPanelRelationPropertyModification";

export interface Product {
  id: number;
  attributes: {
    createdAt: Date;
    updatedAt: Date;
    publishedAt?: Date;
    description: string;
    active: boolean;
    title: string;
    gallery: { data: Media[] };
    slug: string;
    priceId: string;
  };
}
export interface Product_Plain {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  description: string;
  active: boolean;
  title: string;
  gallery: Media[];
  slug: string;
  priceId: string;
}

export interface Product_NoRelations {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  description: string;
  active: boolean;
  title: string;
  gallery: number[];
  slug: string;
  priceId: string;
}

export interface Product_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  description: string;
  active: boolean;
  title: string;
  gallery: AdminPanelRelationPropertyModification<Media>[];
  slug: string;
  priceId: string;
}